[MNT] remove `scikit-learn` as a dependency (#171)
This PR removes `scikit-learn` as a dependency, including the remaining
loci of coupling.

* in `BaseObject`, `scikit-learn` was still used in `clone`. This has
been replaced with a constructor call, and checks that the tests expect.
A config has been added that allows to skip these checks.
* in various tests, `scikit-learn` `clone` was still imported, this is
useful for checking compatibility. Here, the imports have been isolated
inside the tests, so the import is executed only in CI when the `dev`
dependency set is installed. In a case where not, the tests are not
executed.
* `BaseObject` was still inheriting from `sklearn` `BaseEstimator`, but
this should be unnecessary as all has been overridden now. This was
removed, and tests for MRO adjusted.
* `joblib` has been imported in the tests, in a check for parameter
validity. Now, `joblib` is imported, and the check executed only if
`joblib` is present in the environment.