[MNT] resolve dependency conflict in `docs` requirements (#173)
the `skbase` documentation currently does not build due to conflicting
requirements in the `docs` dependency set.

The error is this:
```
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
sphinx-rtd-theme 1.2.0 requires docutils<0.19, but you have docutils 0.19 which is incompatible.
```

This PR resolves the problem by copy-pasting the dependency set from
`sktime`.