[ENH] retain config after `clone`, add config to configure whether to clone config (#257)
This PR ensures that set config fields are cloned, when using `clone`.

This behaviour can be reverted to current via a new config
`clone_config`.

Related: https://github.com/sktime/sktime/pull/5676

Treated as a bugfix as the behaviour it fixes is unexpected.

An example would be, (a) first configuring an estimator via
`set_config`, e.g., parallelization, (b) then wrapping it in a
composite. This procedure would turn off an configs in the internally
cloned estimator.