[ENH] option to return `BaseObject.get_param_names` in the same order as in the `__init__` (#335)
Currently, there is no easy method to return a `BaseObjects`'s parameter
name in the same order as they appear in the `__init__`, which is useful
in various situations. This is because `get_param_names` does retrieve
them, but sorts them alphabetically before returning.

The default should be not sorting, as sorting is an easy operation for
the user, whereas retreving the parameters manually is hard.

This PR therefore introduces a new bool parameter `sorted` to
`get_param_names` which allows the user to specify whether they want the
name sorted.