[BUG] ensure `all_objects` handles decorators properly (#418)
This PR fixes #341 where decorated objects - functions or classes -
would not be retrieved by `all_objects`.

The problem was the check for the containing module of the retreived
object - where the check ends up mistaking the module containing the
decorator for the module containing the actual object.

A test for the bug is included - there are already decorated functions
present in the code base which are now retrieved. This is checked
through the default retrieval checks.