[BUG] regression test certifying for fix #115 (#460)
#### Reference Issues/PRs

Fixes #112 by adding a test


#### What does this implement/fix? Explain your changes.

This does not fix anything, since the bug is apparently already fixed by
#115. But it adds a minimal test.

The test creates a dummy module to test the behaviour of `all_objects`.
This dummy module might be useful for testing, because it is independent
from the existing code. I verified, that the test fails (in an expected
way) before #115 was merged (while all other tests passed). So this test
should give some additional safety in the future.