[ENH] move `git_diff` utility from `sktime` to `skbase` (#471)
#### Reference Issues/PRs

Fixes https://github.com/sktime/skpro/issues/95 after moving git_diff
from sktime and skpro to skbase.

#### What does this implement/fix? Explain your changes.

This PR adds git_diff from utils.git_diff to skbase which is a refactor
because the same code is used in both skpro and sktime, This PR adds
git_diff in skbase and will import this git_diff in skpro and sktime.
The git_diff files and related tests will be removed from skpro and
sktime and will be added here.